ex03/requirements.txt
python-dotenv
langchain
openai
langchain-openai
langchainhub
youtube-search
ollama
langchain-community
cohere==4.46
